More work on neighbor queries, fixed a few bugs.
- Integrated first draft of neighbor management.
  - count() and foreach() tested and working well.
  - kin(), nonkin() implemented but not tested yet.
  - aggregate(), propagate() not implemented yet.

- Added function to call a closure up to completion.

- Made the robot id a uint16_t for more efficient marshalling.

- Fixed bugs:
  - The PUSHL and UNM opcodes were not managed by the VM
  - swarm.exec() was not actually exeucting the closure (only calling it)
  - The parser was not putting global symbols in the base symbol table

+++ buzzvstig.h
@@ -17,7 +17,7 @@ extern "C" {
       /* The timestamp (Lamport clock) */
       uint16_t timestamp;
       /* The robot id */
-      uint32_t robot;
+      uint16_t robot;
    };
    typedef struct buzzvstig_elem_s* buzzvstig_elem_t;
 
@@ -40,7 +40,7 @@ extern "C" {
     */
    extern buzzvstig_elem_t buzzvstig_elem_new(buzzobj_t data,
                                               uint16_t timestamp,
-                                              uint32_t robot);
+                                              uint16_t robot);
 
    /*
     * Clones a virtual stigmergy entry.
